AI-CV-App/requirements.txt
streamlit
openai
fpdf
python-docx
